Make draft invoice generator configurable
